Attempt resolving more warnings.
* initialise in correct order
* initialise members
* macros are evil, use values
* do not calculate stuff with enums
* cast correct type or change function members
* do not try to compare types with enum values

+++ src/world/Management/Guild/Guild.cpp
@@ -301,7 +301,7 @@ void Guild::handleRoster(WorldSession* session)
             << uint32_t(itr->second->getZoneId());
 
         if (!itr->second->getFlags())
-            data << float(float(::time(nullptr) - itr->second->getLogoutTime()) / DAY);
+            data << float(float(::time(nullptr) - itr->second->getLogoutTime()) / static_cast<uint64_t>(DAY));
 
         data << itr->second->getPublicNote();
